CC/requirements.txt
web3>=6.0.0,<7.0.0
solders>=0.10.0,<1.0.0
base58>=2.1.1,<3.0.0
decimal>=1.1
pyyaml>=6.0.1,<7.0.0
requests>=2.31.0,<3.0.0
streamlit>=1.24.0,<2.0.0
cryptography>=41.0.0,<42.0.0
pandas>=2.0.0,<3.0.0
plotly>=5.0.0,<6.0.0
